Update CMakeLists.txt to help Linux find Irrlicht
The lines that set the CH_IRRLICHTDIR seem to be tailored towards Windows. I moved these lines into the Windows (and Darwin) branch so that it doesn't mess up the Linux branch's search for Irrlicht.

The code inside of the Linux branch was changed so that CMake knows where to look for Irrlicht, assuming it was installed by a package manager (if not, the user will have to set the location manually).