Clean up prototypes for the ChCollisionModel::Add***() functions.
Modified the prototypes of all functions used to specify collision
geometry so that they take const refs to the position and orientation
of the contact shapes (instead of pointers to a ChVector and ChMatrix33,
respectively).

This cleans up the code and addresses compilation warnings regarding
'taking address of temporary' (gcc) or 'class rvalue used as lvalue' (cl).