Enable full RPATH linking for installed executables
By default, CMake links executables with full RPATH only in the build tree and
clears the RPATH before installing targets.  This means that, by default, installed
executables will not find the shared libraries (unless these are installed in a default
location, such as /usr/local/lib).

With this change (lifted directly from the CMake wiki), executables are relinked with
the install RPATH.  This allows running the installed executables regardless of the
installation location and without having to change LD_LIBRARY_PATH.