Properly account for the offset and orientation of a visual asset.
A ChVisualization asset defines a transform from the parent (a node in the case of the Irrlicht interface) to the asset.  This must be concatenated with the transform from the asset to the geometry.  Previously, only the latter was accounted for.

Note: I only fixed the sphere, cylinder, and box visual assets.  I still need to look at ChObjShapeFile and ChTriangleMeshShape.