Add global free functions for dealing with the Chrono data.
While this set of changes do not address the issue of where the data/ directory lives, they offer the following:
(1) it isolates the hard-coded default location (namely "../data"), relative to the location of the executable programs, to a single place (in ChGlobal.cpp)
(2) it allows external projects that depend on Chrono to access the default Chrono data files without having to copy the data/ directory.  Instead, simply invoke SetChronoDataPath() to specify where the Chrono data directory is (for example, in the Chrono source tree or the Chrono install tree)

The three new functions are as follows:
(1) SetChronoDataPath(): specifies the location of the Chrono data/ directory. The default is "../data/".  Note that this function is not thread-safe.
(2) GetChronoDataPath(): retrieves the location of the Chrono data/ directory (thread-safe)
(3) GetChronoDataFile(): obtains the complete filename given the name of a file or subdirectory relative to the location of the Chrono data directory (thread-safe)