Use ChLinkSpringCB for the spring and shock force elements in a ChDoubleWishbone suspension.
This provides greater flexibility and allows for nonlinear spring and/or dampers.
By default we use two internal functor classes which implement linear springs and dampers, respectively, using the coefficients specified through getSpringCoefficient() and getDampingCoefficient(); this is equivalent to the previous behavior.  Optionally, a derived class can use a custom implementation for either force element.  This must be specified by overriding useNonlinearSpring() and/or useNonlinearShock() to return true and providing custom functor classes through getSpringForceCallback() and getShockForceCallback(), respectively.