Fix incorrect calculations in ChPacejkaTire::getKAG_from_State().
Although this function is currently not used anywhere, it should be used as a model to fix all similar calculations throughout ChPacejkaTire!

Note that the slip angle, camber angle, and longitudinal and lateral velocities are all calculated in a frame with Z up, X forward, and Y to the left, such that a positive longitudinal velocity and positive wheel angular speed correspond to forward motion.  Because of the current chrono-T reference frames (with X pointing towards the rear), this requires some explicit conversions (which will be removed once we switch to an ISO reference frame).