When initializing a suspension, pass the body to which the tierods should be attached.
For a steerable suspension, this should be the steering link body of a steering subsystem.  Otherwise, this is the chassis body.
It is now important to always initialize the steering subsystem *before* the suspension subsystems.

(Note: the tierods for the various suspension templates do not connect to these bodies yet.)