Modify implementation of collision families to use bit masks.
For consistency with the bullet collision system, we use (signed) shorts to hold the collision family (encoded in 'family_group' as a power of 2, with the position of the set bit indicating the family) and the collision mask (encoded in 'family_mask' with a set bit indicating collision with the corresponding family).

This limits the number of collision families to one of the following 15 values: {0, 1, ..., 14}.  Note that we use signed shorts so we must be careful with the sign bit.

A "no collision" relationship is commutative but not transitive.  Note that, unless explicitly changed, the bit in a model's mask corresponding to the model's own family is set.  As such, this should not be relied upon to ignore self-collisions.

Collision family information in this format is now saved and retrieved from checkpoint files.  This invalidates previous checkpoint files which should be regenerated.