Add new joint ChLinkUniversal, with the kinematics commonly associated with a 'universal' joint.
TODO: implement calculation of reaction forces/torques.

Note that the link-lock formulation does not allow proper modeling of a universal joint (in the sense of the definition commonly used, at least in the US).  The resulting 2 rotational degrees of freedom joint with the link-lock formulation (the current ChLinkLockUniversal) may actually be a constant-velocity joint.  I need to check this;  if this is indeed the case, we will rename this joint to something like 'ChLinkLockConstantVelocity'; otherwise, we will simply remove it.