Disable parallel processing of the loop over the otherphysics list in ChSystem::LCPprepare_load.
The problem is with items such as ChShaftsTorqueBase which add torques to other items' variables.  As such, there is potential contention with the various VariablesFbLoadForces and VariablesFbIncrementMq functions.
While it would be possible to protect the calls to the above two functions in additional critical sections, more experimentation is needed to decide if it's worth it.  In the meantime, the loop is processed sequentially.