Small fixes for APGD
In certain cases the L value was going to zero causing the solver to iterate indefinitely.
The gamma values were overwritten with an empty vector which would cause the code to crash in debug mode.
If L is zero I terminate the solve because I don't think the solver can do anything... This may need to be looked at further