This change compiles and links on Linux/OSX but will cause simulations to segfault immediately. Looking into a better fix currently
Revert "Trouble... easylogging.h is included in the ChParallelDefines.h , so it is included in most sources that deal with chrono paralel,including the open gl viewer library and the chrono utils library, and all demos in chrono parallel. The fact is that, at leas in windows, this gives a ton of 'undefined symbols' when linking, unless one adds the INITIALIZE_EASYLOGGING macro in some cpp. I did so for the openGL viewer and the chrono utils library, but still it should be added to all demos. Maybe there is a less 'invasive' fix?"

This reverts commit 0903687214b9fc355ea87c8d2b6159a0886c432e.