Disable update of visualization assets where appropriate.
During the step integration, the Update() function is called twice, once at the current state and again after the state was advanced to the new time.
During the former, there is no need to update visualization assets.
For models with complicated assets, this can significantly reduce the cost of an integration step (up to 2x)

TODO: revisit all other places where updates are performed and disable asset updates if appropriate.