Add skeleton for RSD-based anti-roll bar template.
This class implements a parameterized template for an anti-roll bar subsystem modeled with two arm bodies connected through a rotational spring-damper.  Each arm is connected through a droplink (modeled as distance constraints) to bodies of an associated suspension subsystem.   Finally, one of the arm bodies is connected to the chassis through a revolute joint.