Rename the folder for the utility library in unit_VEHICLE to `chrono_vehicle_utils`.
This change is required for programs that use both unit_VEHICLE and unit_PARALLEL (to distinguish between the utility libraries in the two units).
NOTE:  the classes and functions in these utility libraries will be eventually combined and/or distributed in the appropriate Chrono units.  At that point, the two libraries ChronoEngine_VehicleUtils and ChronoEngine_ParallelUtils will be obsolete.