Add virtual method ChSystem::NewBody.
This method, which must be overriden by derived classes, creates and returns a pointer to a new body.  This body is created with a contact method consistent with the type of the ChSystem and with the collision system currently associated with the ChSystem.  Note that the body is not attached to the system.

This change facilitates writing generic utility functions that create bodies.