Add utility Chrono::Vehicle classes to implement a steering PID controller.
The core functionality is implemented in the abstract base class ChSteeringController
- implements basic functionality to control the error between the location of a sentinel point (a point at a look-ahead distance in front of the vehicle) and the current target point.
- a derived class must implement the CalcCurrentTarget method
- error derivative is estimated using backward finite differences
- error integral is estimated using the trapezoidal integration scheme, with a step-size

The class ChPathSteeringController is a derived class which sets the target location as the closest point on a Bezier path to the sentinel point.
- it uses a Bezier path tracker object to robustly calculate the target location

Additional derived classes could set the target location from other sources (e.g. by interfacing with a camera sensor model).