Fix for running unit tests that use the Chrono data directory.
Because we use a relative path to the Chrono data directory, we must ensure that CTest executes the tests in the appropriate directory.  This requires specifying the WORKING_DIRECTORY test property.  However, matters are further complicated by the fact that VS places the executables in a subdirectory of ${EXECUTABLE_OUTPUT_PATH} (depending on the build configuration).