Allow specifying the path to the Chrono::Vehicle data directory independent of that for the Chrono data.
The default setting is still for the Chrono::Vehicle data directory to be a subdirectory 'vehicle/' of the Chrono data directory.
However, the user can override this through a call to chrono::vehicle::SetDataPath().
The Chrono package configuration file now exports a variable CHRONO_VEHICLE_DATA_DIR (with value "${CHRONO_DATA_DIR}vehicle/") which can be used in external projects to find the path to the default Chrono::Vehicle data directory.