HHT: switch to relative + absolute NR convergence test (in ACCELERATION mode).
NR is declared converged if either the L2 norm of the residual is below the absolute tolerance or the WRMS norm of the update is less than 1.
The WRMS norm test is a relative + absolute tolerance test:
    ||v||_WRMS = sqrt{ sum{(v_i * w_i}^2 / N }
where the weights are calculated as
    w_i = 1 / (reltol * |v0_i| + abstol)
with v0 the current solution (i.e., the solution at previous step).
WRMS norms for the update in the accelerations and Lagrange multipliers are calculated separately and use possibly different absolute tolerances.