Fix bug in promoting raw pointer to shared_ptr (use empty destructor).
This issue occurs whenever a raw pointer must be promoted to shared_ptr (e.g. to be passed to some function) but that pointer is already managed through some other shared_ptr.  Without explicitly specifying a custom empty (no-op) destructor, the raw pointer would be deleted when the local (temporary) shared_ptr goes out of scope.

(Note: with the previous ChSharedPtr implementation, this was done by explicitly increasing the reference count of the wrapped ChShared object)