Disable option for setting the precision in Bullet.
When Bullet is compiled in double precision, some collisions are missed (with MSVC, this only happens in Release builds but not in Debug builds).
For now, until this issue is resolved, we always build Bullet in single precision.
Note however that Bullet has some other issues when compiled in single precision (e.g. some tests for degenerate tetrahedrons are not correct).