Implement remaining ChCollidable virtual methods for ChContactTriangleXYZ.
The treatment here is slightly different than for other ChCollidable objects because the contact triangles are always referenced to the absolute frame.  To address this, the implementation is based on assuming that the barycentric coordinates of contact points do not change with changes in the object state (i.e., changes in the coordinates of its three vertices).