Report tire forces differently for the case of co-simulated tires.
Tires that have actual physical models (such as rigid tires or FEA-based tires) must report zero force and moment (because the tire forces on the associated wheel body are applied implicitly, either as contact forces or as link reaction forces).  However, when the tire is co-simulated, GetTireForce() must return the actual force to be applied to the associated wheel body.

Other tire models (namely those modeled as force elements, such as Pacejka, Fiala, Lugre) ignore this flag and always calculate and report the actual tire force to be applied to the wheel body.