Add (pure) virtual method ChElementBase::GetNodeNdofs().
This function must return the number of coordinates associated with the n-th node that are used by the current element.
Note that this may be different from the value obtained through GetNode(n)->Get_ndof_w().
For all current elements, this function will return the same value for all the nodes (however, future elements may use combinations of different nodes).