Fix functions for calculating element contributions to nodal generalized forces.
This allows a given element to use derived node classes (e.g. a ChNodeFEAxyzD instead of ChNodeFEAxyz) where appropriate to permit different element types to share nodes in the same mesh.
The fix involves asking a given element for the number of nodal coordinates it affects for a given node (instead of asking the node itself for its total number of coordinates) while including the contributions to the nodal generalized forces.

Limitations:  an element always adds its contribution to the first coordinates of its adjacent nodes.