Create a library of Chrono::Vehicle models.
The directory hierarchy src/models/ is meant to collect various models (corresponding to the different modules in Chrono) that could be reused (for demos, tests, or third-party projects).  Each set of such models should be encapsulated in a separate library.

Currently, the ChronoModels_vehicle library contains the models for the 'hmmwv' and 'm113' vehicles (code was moved from demos/vehicle/hmmwv and demos/vehicle/m113).