Add new FEA node.
This node may be used for adding a new position vector gradient variable in a node. For example, it can be used to model an element whose nodes have 2 position vector gradients in addition to the nodal position vector. In addition, the extra vector (DD vector) can also be used to represent other type of vector, e.g. a curvature vector.