Add 3-node continuum-based, fully 3D ANCF beam element.
This element models the internal forces of a beam using a continuum based approach. Poisson locking is avoided by integrating Poisson terms over the beam centerline. This element captures, extension, bending, torsion, and shear deformation. Additionally, the user can select to account for Poisson effect. The initial configuration of the FEA system is automatically considered as the reference. Isotropic and orthotropic materials can be selected. Further details on the features of this ANCF beam element may be found in Nachbagauer, Gruber, and Gerstmayr, "Structural and Continuum Mechanics Approaches for a 3D Shear Deformable ANCF Beam Finite Element:  Application to Static and Linearized Dynamic Examples", Journal of Computational and Nonlinear Dynamics, 2013, April 2013, Vol. 8, 021004.