Remove redundant constraints in double-pin track assembly.
To ensure that the (open) chain of track shoes is not over-constrained, use spherical joints between connector bodies of a track shoe and the shoe body of the neighbor.
To ensure that the (closed) chain of track shoes is not over-constrained, replace one set of spherical joints with a point-line joints.
To accommodate an arbitrary number of track shoes, we special-case the connection between the first two track shoes.