Support composite inertia properties in JSON specification for a Chrono::Vehicle chassis.
The inertia properties of the chassis rigid body is calculated for a collection of sub-systems (all assumed rigidly attached to the chassis body).

Each sub-system is specified through a centroidal frame (relative to the chassis reference frame), a mass, moments of inertia (w.r.t. the sub-component's centroidal frame), and product of inertia (w.r.t. the sub-component's centroidal frame).  Optionally, a sub-component can be marked as a "void", in which case its inertia properties are subtracted from that of the composite.  Note that products of inertia are assumed to include the minus sign; in other words, the provided 3-dimensional vector should contain [ -int{xy}dm   -int{xz}dm   -int{yz}dm ]