Fix CMake configuration script for Chrono::Python module.
These rewrite of the CMakeLists.txt script works properly on both Windows and Linux systems.
The SWIG-generated files (*.py) and module libraries (*.pyd on Windows, *.so on Linux) are collected into a single directory in the BUILD tree (so that they can be used directly from there).
On installation, these files are collected into a single directory in the INSTALL tree (different on Windows and Linux).
Use of the Chrono::Python modules requires the user to set the PYTHONPATH environment variable.