Fixes POSIX threads implementation on macOS/OSX
Chrono Parallel relies on an optional feature (memory-based semaphores)
for its implementation of ChThreadsPOSIX. This feature is not available
in Mac (Darwin) based kernels, and causes a "spinning wheel of death" to
appear when demos using it are executed. (tested on macOS Sierra 10.12.3)

This bug has been present for an unknown length of time, and has been
indicated on recent builds with the warning -Wdeprecated-declarations.

Implemented with platform-specific defines, the now-working implementation
uses POSIX named semaphores. This implementation may actually be cross-
platform for UNIX-like systems because it relies only on required
pthreads features, but for the purposes of this feature branch, it will be
only on Apple.