Modifies CMake Config to export FMA build flags
This allows projects that link against Chrono to read the '-mfma' flag
required for gcc and clang to use FMA libraries. For example, this
allows the chrono-projects repository to build against chrono with
chrono-parallel and chrono-opengl enabled, using out-of-box brew
versions of GLM. Without this fix the build will not use FMA and fail.