Change law for combining rolling and spinning friction in Chrono::Parallel.
For consistency with sequential Chrono, use the minimum of the two material values to set the rolling and spinning friction values for a contacting pair.
Note that the combining law for sliding coefficient of friction was left unchanged (and different from the one in Chrono, where the minimum is used).