Adds .cuh headers and unit_tests to install target
This commit adds all .cuh headers in src/chrono_parallel to the chrono
install target if USE_PARALLEL_CUDA is enabled. This is a workaround for
various chrono headers including CUDA headers that are currently not
included in the install directory. It also adds unit tests back to the
install directory to allow testing of installation integrity