Introduce a class to represent a rigid-body chassis template.
This class will manage visualization and contact shapes for a chassis modeled as a single rigid body.
Note that this class does not resolve the pure virtual methods of ChChassis so it cannot be instantiated.

The rigid body member of ChChassis represent, in general, the frame of the chassis (so, for example, a chassis template that adds flexible panels for the rest of the chassis could be implemented in the future)