Rigid CB Track Development In Progress - INCOMPLETE
Currently the code started with the double pin track model and is slowly being morphed and tested to switch it over to the new continuous band track model.  There are large amounts of the code that need to be cleaned up as most of the code is still in the process of being switched over.
1) Sprocket geometry is showing the new CB profile (hard coded profile parameters right now), but the collision is still the double-pin
2) The new CB assembly algorithm has been roughed in and tested, but the original double pin code needs to be cleaned out and removed.