Completed initial custom collision for the sprocket and rigid continuous band track belt
Code clean up, including lots of additional comments
Currently the code uses revolute joints to connect the CB track belt segments.  Still working on the bushing elements.
The exact size of the belt and sprocket are TBD.  There is currently a scaling factor on the relevant parameters to scale the belt and sprocket closer the right size for this model