Fix bugs in vehicle lateral controllers (for non-ISO world frames).
A recent change modified the reporting of vehicle orientation so that it is consistent with the current world frame.
However, the path-following lateral controllers (of types "Human", "XT", and "SR") had the built-in assumption that the Y axis of the vehicle rotation matrix is the vehicle's left direction.
The fix is to always perform these calculation using the ISO reference frame and only convert to the current world frame when needed.