Clamp regions for ray-casting to available grid index range.
Note:
For a flat SCM terrain patch, this clamping is only needed if using a visualization mesh (as we would access mesh vertices out of bounds).
The rest of the grid-based SCM algorithm accommodates an infinite grid.
For an SCM terrain patch initialized from a height-filed map, this clamping is also needed in order not to access the heights table out of bounds (this could be relaxed by defining the undeformed terrain height beyond the specified limits to be that at the slosest map boundary).