Add new collision shape: cylindrical shell.
This collision shape represents a sphere-swept cylindrical shell, defined by a radius and halh-length of the skeleton cylinder and a radius of the sweeping sphere.
A cylshell is always aligned with the Y axis.

This collision type defaults to acting like a collision cylinder with inflated radius and half-length in the Bullet collision system and like a rounded cylinder in the Chrono::Parallel collision system.

The main purpose of this collision shape is to allow custom algorithms for certain collision pairs involving a cylshell.  In particular, the cylshell-box collision will be customized both in Bullet and Chrono::Parallel.