Modify test for 64-bit cl compiler in FindSSE.cmake.
The previous test (using CMAKE_CL_64) is not working with the latest version of Visual Studio.
Replace to a test based on CMAKE_SIZEOF_VOID_P.

Indeed, when compiling for x64 with VS, the flag "/arch:SSE2" is not needed anymore (it's automatic) and its explicit definition resulted in many cl warnings.