Updating the new ANCF elements so that the fast ANCF specific generalized force due to gravity method is now efficiently calculated at every internal force call.
This change accounts for cases where the acceleration due to gravity for the containing system changes between time steps (for each during a unit test with a gravity load).
Prior to this, it was assumed that the acceleration due to gravity for the containing system was constant and it was set prior to the precalculation phase for the elements.