Fix bug in ChTriangleMeshConnected::RepairDuplicateVertexes.
One cannot modify the normal, UV, and color face indices because these are not guaranteed to be the exact same ones as the vertex indices!
Instead, they must continue to point in the original lists of normals, UVs, and colors, respectively.
This function is intended to only remove vertices that are geometrically close to each other.  And not to modify other mesh properties!