Fix order of synchronization in wheeled vehicle FMU.
It is critical to synchronize the wheels of the vehicle FMU *after* synchronizing the vehicle system itself.
This is because the vehicle FMU does not include tires and, as such, tire/terrain forces are explicitly applied to the wheel bodies.
If this is based on using the wheel body force accumulators, those are cleared during vehicle (suspension) synchronization which therefore must occur *before* tire/terrain forces are included.