Add mechanism to disable self-collisions for a ChContactSurface (node cloud or trimesh).
By default, self-collision are allowed and included in collision detection.
If appropriate for a given simulation, these can be disabled by specifying a collision family for the contact surface.
If a family is provided, all collision models contained in the contact surface are assigned to that family and collisions are disabled within the family.

(cherry picked from commit 900bc3b67144213309b633f4184094bf8236a0fc)