Add ChTire option to set the collision family for the tire contact model(s).
Concrete tire models that rely on the Chrono underlying collision detection for tire-terrain interaction may use this setting.
One possible use if to disable self-collision (e.g., for models that use a ChContactSurface); this can be done by assigning all component collision models to the same family and disabling collisions with that family.
The tire collision family (default: 13) can be set through the ChTire::SetContactSurfaceType() function.