Add a generic FSI interface, agnostic of the particular fluid solver type.
This generic FSI interface intermediates the data exchange betwen the associated Chrono multibody system and fluid solver using Chrono data types (i.e., independent of the fluid solver).
A particular fluid solver may be paired with a custom FSI interface (as is the case with the current SPH-based fluid solver).
The custom FSI interface is potentially more performant than the generic one, as it can work directly with the fluid solver memory and thus save an additional data copy that is inherent in the generic FSI interface.