define Euler beam quaternion upon building
This is done for simplicity but I also think there might be a bug in SetupInitial().
When the beam is created from 2 nodes, or 1 node and 1 position, the end nodes may
have different Y axes, and there is no reason why this Y axis should correspond to
that of the beam element since those are determined by who created these nodes.